Code simplification

+++ src/main/java/com/solvd/dao/IAccesorieDao.java
@@ -3,6 +3,6 @@ package com.solvd.dao;
 
 import com.solvd.pojos.Accesorie;
 
-public interface IAccesorieDao extends IDao <Accesorie, Long> {
+public interface IAccesorieDao extends IDao <Accesorie, Integer> {
 
 }
