+++ src/main/java/com/solvd/dao/IClientDao.java
@@ -2,5 +2,5 @@ package com.solvd.dao;
 
 import com.solvd.pojos.Client;
 
-public interface IClientDao extends IDao <Client, Long>{
+public interface IClientDao extends IDao <Client, Integer>{
 }
