+++ src/main/java/com/solvd/dao/IDoctorDao.java
@@ -2,5 +2,5 @@ package com.solvd.dao;
 
 import com.solvd.pojos.Doctor;
 
-public interface IDoctorDao extends IDao <Doctor, Long>{
+public interface IDoctorDao extends IDao <Doctor, Integer>{
 }
