+++ src/main/java/com/solvd/dao/IFoodDao.java
@@ -2,5 +2,5 @@ package com.solvd.dao;
 
 import com.solvd.pojos.Food;
 
-public interface IFoodDao extends IDao <Food, Long> {
+public interface IFoodDao extends IDao <Food, Integer> {
 }
