+++ src/main/java/com/solvd/dao/IMedicineDao.java
@@ -2,5 +2,5 @@ package com.solvd.dao;
 
 import com.solvd.pojos.Medicine;
 
-public interface IMedicineDao extends IDao <Medicine, Long> {
+public interface IMedicineDao extends IDao <Medicine, Integer> {
 }
