+++ src/main/java/com/solvd/dao/IPaymentDao.java
@@ -2,5 +2,5 @@ package com.solvd.dao;
 
 import com.solvd.pojos.Payment;
 
-public interface IPaymentDao extends IDao <Payment, Long>{
+public interface IPaymentDao extends IDao <Payment, Integer>{
 }
