+++ src/main/java/com/solvd/dao/ITreatmentDao.java
@@ -2,5 +2,5 @@ package com.solvd.dao;
 
 import com.solvd.pojos.Treatment;
 
-public interface ITreatmentDao extends IDao <Treatment, Long>{
+public interface ITreatmentDao extends IDao <Treatment, Integer>{
 }
