+++ src/main/java/com/solvd/dao/IVeterinaryDao.java
@@ -2,5 +2,5 @@ package com.solvd.dao;
 
 import com.solvd.pojos.Veterinary;
 
-public interface IVeterinaryDao extends IDao <Veterinary, Long> {
+public interface IVeterinaryDao extends IDao <Veterinary, Integer> {
 }
